Add a raw format to sources to skip extraction for single-file downloads
This allows for specifying the format 'raw' as well as a desired filename,
which will download the file to the regular archive directory and then
copy it over to the the destination 'subdir', renaming the file to
'filename'. Example:

tools:
- name: ovmf
  sources:
    subdir: 'misc'
    url: 'https://your.url/path/to/ovmf-pure-efi.fd'
    format: 'raw'
    filename: 'OVMF.fd'

This will download the file 'ovmf-pure-efi.fd' into your archive
directory. Then, it will be copied over to 'misc/ovmf/OVMF.fd'.