Allow for patching archives, i.e. without git or hg
This allows for patching of archives downloaded, e.g. as tar.gz. This
requires the patch to be formatted with 'diff -Naur', which also makes
it similar to git's patches. The path strip depth ('patch -p<value>')
defaults to 1, but can be configured with the 'patch-path-strip' option
in the 'source' scope.