Add xbstrap-pipeline script
xbstrap-pipeline is an interface to xbstrap that is better
suited for unattended/automated builds (such as CI builds).
In contrast to xbstrap, it is not centered upon build steps.
Instead, it divides the build into jobs that produce build
artifacts (i.e., tools and packages).

Jobs can be scheduled independently from each other among
multiple servers as long as dependencies are respected. The
scheduling mechanism itself is out of scope of xbstrap.